Use /proc/sys/kernel/pid_max to better keep the report tabular.  This file has existed since before the Linux kernel was under git in 2005.
Note that due to the module global nature of `fAdd`, this does incur an
open/read/close time penalty even when unneeded, such as in procs.find()
where we attend performance, but this is probably still small enough
compared to likely hundreds of /proc/*/ files to read to not worry about
much.  Even so, leave a note.