Take the "set union" of all labels for a given process
Sometimes a PID appears as input more than once, but with a different
set of labels each time.  E.g., pd ab1234 a1234.  This PR essentially
does a set union on the labels, so that PID 1234 would have labels "ab".
(The previous behavior would keep the set of labels that appear with the
last occurrence of the PID; here, just "a".)