Allow for longer external labels with `string` -> `seq[string]`.
Make column header "LABEL", giving room for 2 length-2 labels with one
delimiter, or maybe a 1,3 or a 3*1 triplet before starting to misalign
the text table.

Also simplify `labels[] = labels & l` to just `labels.add`.

While we need a ':' delimiter for formatting, we do not for input.  You
just use more than one argv slot to assign multiple labels to a given
PID, but I think for the intended use of external PID typing that is ok.
For example, for 3 existing PIDs, this seems to work:
```
$ procs d -f'%l %p' aa741 742 bx743 ac741
LABEL   PID
aa:ac   741
        742
bx      743
$ procs d -f'%l %p' 741 742 743
LABEL   PID
        741
        742
        743
```
Right now we do not use a `HashSet` partly since under 4 labels are
expected and partly to ensure order.  In the above example it will be
"aa:ac" as long as aa741 precedes ac741 in the argv, never "ac:aa".
This may matter for matching regular expressions in `pcr_l` which are
sensitive to textual order without all permutations present (which is
surely possible, but also surely annoying).