pidfs's started in 2019 and glibc started supporting in 2022.  So, finally do the old XXX work item which was never much work/perturbation to anything as this patch shows.
Since there is no `nice` interface in Linux taking pidfd's, we just add
a new `kill` overload taking 3 args and a new argument to `act` that is
-1 if there is no pidfd available & change `kill` -> `newOvrLd |
pidfd_send_signal` in appropriate places.  Pid is a type alias for cint
which is the file descriptor type - so type-based overload isn't enough.

Inside `procs find` we just pair up pList maintenance with a new fList
for the pidfds and pass the right things to the new version of `act` &
kill in kill-ing contexts (printing needs to assume the same decimal PID
means the same thing anyway).

Oh, and in the acting-ASAP mode, we simply use the vanilla / raw PID
directly. It only costs time not saving any races to do otherwise.