allow sorting by 'm'; include meminfo in cpio archive by default (#5)
Currently, 'm' (%MEM) works in the format spec, but not in the sort
spec.  It's really the same as 'R', but it's nice to be able to use the
same letter for both format and sort.  This PR adds 'm' as an alias for
'R'.  It also includes meminfo in the parc scripts, which is required
for %MEM.